% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\docType{data}
\name{categories}
\alias{categories}
\title{Categories}
\format{
A data frame with 8 rows and 4 variables:
\describe{
\item{category_id}{Category ID, primary key}
\item{category_name}{Category name}
\item{description}{Category description}
\item{picture}{Picture (binary data)}
}
}
\source{
Northwind sample database
}
\usage{
categories
}
\description{
Product categories for organizing the Northwind product catalog.
}
\examples{
  categories
  
  # Products per category
  library(dplyr)
  categories \%>\%
    left_join(products, by = "category_id") \%>\%
    count(category_name, sort = TRUE, name = "product_count")
  
  # Category details
  categories \%>\%
    select(category_name, description)
}
\keyword{datasets}
