% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Northwind.R
\docType{data}
\name{Customers}
\alias{Customers}
\title{Northwind Customer data table}
\format{A dataframe with 29 rows and 12 variables:
\itemize{
\item ID.
\item Company. name of the customer
\item LastName. last name of point of contact at the customer
\item FirstName. first name of point of contact at the customer
\item JobTitle. title of point of contact at the customer
\item BusinessPhone. formatted as "(123)456-7890"
\item FaxNumber. formatted as "(123)456-7890"
\item Address.
\item City.
\item StateProvince. Two digit abbreviation
\item ZIPPostalCode.
\item CountryRegion.
}}
\usage{
data(Northwind)
}
\description{
This data table is part of the freely downloadable Northwind sample database.
They are useful datasets for demonstrating such things as SQL queries and
business data analytics. Loading the Northwind dataset will place this and other
related datanto your environment.
}
\keyword{data}

