% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Northwind.R
\docType{data}
\name{Employees}
\alias{Employees}
\title{Northwind Employees data table}
\format{A dataframe with 9 rows and 15 variables:
\itemize{
\item ID.
\item Company. name of the customer
\item LastName.
\item FirstName.
\item EmailAddress
\item JobTitle.
\item BusinessPhone. formatted as "(123)456-7890"
\item HomePhone. formatted as "(123)456-7890"
\item FaxNumber. formatted as "(123)456-7890"
\item Address.
\item City.
\item StateProvince. Two digit abbreviation
\item ZIPPostalCode.
\item CountryRegion.
\item Wibpage. A fictitous URL
\item Notes.
}}
\usage{
data(Northwind)
}
\description{
This data table is part of the freely downloadable Northwind sample database.
They are useful datasets for demonstrating such things as SQL queries and
business data analytics. Loading the Northwind dataset will place this and other
related datanto your environment.
}
\keyword{data}

