% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Northwind.R
\docType{data}
\name{InventoryTransactions}
\alias{InventoryTransactions}
\title{Northwind Inventory Transactions data table}
\format{A dataframe with 102 rows and 6 variables:
\itemize{
\item TransactionID.
\item TransactionType.
\item TrnsactionCreatedDate.
\item TransactionModifiedDate.
\item ProductID.
\item Quantity.
}}
\usage{
data(Northwind)
}
\description{
This data table is part of the freely downloadable Northwind sample database.
They are useful datasets for demonstrating such things as SQL queries and
business data analytics. Loading the Northwind dataset will place this and other
related datanto your environment.
}
\keyword{data}

