% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InventoryTransactionTypes.R
\docType{data}
\name{InventoryTransactionTypes}
\alias{InventoryTransactionTypes}
\title{Northwind Inventory Transactions Type data table}
\format{A dataframe with 4 rows and 2 variables:
\itemize{
\item ID.
\item TypeName.
}}
\usage{
data(Northwind)
}
\description{
This data table is part of the freely downloadable Northwind sample database.
They are useful datasets for demonstrating such things as SQL queries and
business data analytics. Loading the Northwind dataset will place this and other
related datanto your environment.
}
\keyword{data}
