% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/northwind_list.R
\name{northwind_list}
\alias{northwind_list}
\title{List of Available Northwind Tables}
\usage{
northwind_list()
}
\value{
A tibble with two columns: \code{table} and \code{description}
}
\description{
Returns a tibble with the names and descriptions of the available tables
included in the \code{northwind} package.
}
\examples{
northwind_list()
}
