% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrderDetails.R
\docType{data}
\name{OrderDetails}
\alias{OrderDetails}
\title{Northwind Order Details}
\format{dataframe}
\usage{
data(Northwind)
}
\description{
This data table is part of the freely downloadable Northwind sample database.
They are useful datasets for demonstrating such things as SQL queries and
business data analytics. Loading the Northwind dataset will place this and
other related data into your environment.
}
\keyword{data}
