% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\docType{data}
\name{regions}
\alias{regions}
\title{Regions}
\format{
A data frame with 4 rows and 2 variables:
\describe{
\item{region_id}{Region ID, primary key}
\item{region_description}{Region description/name}
}
}
\source{
Northwind sample database
}
\usage{
regions
}
\description{
Sales regions used to group territories for Northwind's sales organization.
}
\examples{
  regions
  
  # Count territories per region
  library(dplyr)
  regions \%>\%
    left_join(territories, by = "region_id") \%>\%
    count(region_description, name = "territory_count") \%>\%
    arrange(desc(territory_count))
}
\keyword{datasets}
