% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Suppliers.R
\docType{data}
\name{Suppliers}
\alias{Suppliers}
\title{Northwind Suppliers data table}
\format{A dataframe with 9 rows and 15 variables:
\itemize{
\item ID.
\item Company. name of the customer
\item LastName.
\item FirstName.
\item EmailAddress
\item JobTitle.
}}
\usage{
data(Northwind)
}
\description{
This data table is part of the freely downloadable Northwind sample database.
They are useful datasets for demonstrating such things as SQL queries and
business data analytics. Loading the Northwind dataset will place this and other
related datanto your environment.
}
\keyword{data}
