% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/territories.R
\docType{data}
\name{territories}
\alias{territories}
\title{Territories}
\format{
A data frame with 53 rows and 3 variables:
\describe{
\item{territory_id}{Territory ID, primary key}
\item{territory_description}{Territory description/name}
\item{region_id}{Region ID, foreign key to regions table}
}
}
\source{
Northwind sample database
}
\usage{
territories
}
\description{
Sales territories used by Northwind employees for organizing sales activities.
}
\examples{
  territories
  
  # View territories by region
  library(dplyr)
  territories \%>\%
    left_join(regions, by = "region_id") \%>\%
    select(territory_description, region_description) \%>\%
    arrange(region_description, territory_description)
}
\keyword{datasets}
