% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Northwind.R
\docType{package}
\name{Northwind}
\alias{Northwind}
\alias{Northwind-package}
\title{Northwind: a collection of database tables}
\description{
The Northwind package is a collection of database tables from the freely downloadable
Northwind sample database  They are useful datasets for demonstrating such things as
SQL queries and business data analytics. The tables in this package include:
\itemize{
\item Customers
\item Employees
\item InventoryTransactions
\item InventoryTransactionTypes
\item Invoices
\item Orders
\item PurchaseOrderDetails
\item PurchaseOrders
\item Shippers
\item Suppliers
\item
}
}
