% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shippers.R
\docType{data}
\name{shippers}
\alias{shippers}
\title{Shippers}
\format{
A data frame with 3 rows and 3 variables:
\describe{
\item{shipper_id}{Shipper ID, primary key}
\item{company_name}{Company name}
\item{phone}{Phone number}
}
}
\source{
Northwind sample database
}
\usage{
shippers
}
\description{
Shipping companies used by Northwind.
}
\examples{
  shippers
  
  # Shipper usage frequency
  library(dplyr)
  shippers \%>\%
    left_join(orders, by = c("shipper_id" = "ship_via")) \%>\%
    count(company_name, sort = TRUE, name = "orders_shipped")
}
\keyword{datasets}
